+++ generators/app/writing/gruntConfig.js
@@ -40,6 +40,7 @@
 		'preprocess',
 		'prettier',
 		'prompt',
+		'pug',
 		'release',
 		'string-replace',
 		'usemin',